Part 2
This is the Part 2 of my Web Dev project, where i added the use of Css-Styles in addition to the first project where it was the Html. i used the styling to design my website and add features to make it popout and very pleasing, i then made changes to the mistakes i did prior from the navigation bar to also making it responsive to all kinds of screen and placed different kinds of features from word placement to colouring and also linked it to my Css.